% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{ar1g_timedep_sim}
\alias{ar1g_timedep_sim}
\title{Simulate time dependent AR(1) series (using Green's function)}
\usage{
ar1g_timedep_sim(n, sigma = 1, a = 0.2, b = 0.7, phis = NULL)
}
\arguments{
\item{n}{The length of the simulated time series.}

\item{sigma}{The standard deviation of the innovations.}

\item{a}{Intercept in the evolution of the lag-one correlation.}

\item{b}{Slope in the evolution of the lag-one correlation.}

\item{phis}{Numeric of length \code{n}. If evolution of lag-one correlation is 
to be given explicitly this is done here. Overrides \code{a} and \code{b}.}
}
\value{
Returns the simulated time series as a \code{numeric} object.
}
\description{
This function produces samples from a given time dependent AR(1) process using Green's function instead of AR(1) directly.
}
\examples{
\donttest{
n = 600
a=0.5
b=-0.1
phis = seq(from=a,to=a+b,length.out=n)
sims = ar1g_timedep_sim(n,sigma=1,a=a,b=b)
}
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{ar1}
\keyword{simulation}
\keyword{timedep}
