% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{beta2}
\alias{beta2}
\title{Beta1 computation}
\usage{
beta2(Ht, Hs)
}
\arguments{
\item{Ht}{First Hurst exponent}

\item{Hs}{Second Hurst exponent}
}
\value{
Returns the value of the function.
}
\description{
Necessary for computing covariance function of time-varying fGn
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fgn}
\keyword{simulation}
\keyword{timedep}
