% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_mu_ar1}
\alias{c_mu_ar1}
\title{Compute mu in C for AR(1) model}
\usage{
c_mu_ar1(means, z, n, lambda, time)
}
\arguments{
\item{means}{Vector that will be overwritten with the forced response.}

\item{z}{Vector with the shifted and scaled known forcing.}

\item{n}{Integer denoting the dimension of \code{mu}.}

\item{lambda}{Vector describing the lambdas at each sampled time step.}

\item{time}{Vector describing the time steps.}
}
\description{
Void function that gives the forced response for the AR(1) model with 
time-dependent memory in C.
}
\seealso{
\code{\link{mucomputer}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{C}
\keyword{ar1}
\keyword{forcing}
