% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{ch}
\alias{ch}
\title{Time dependent constant}
\usage{
ch(H)
}
\arguments{
\item{H}{Hurst exponent}
}
\value{
Returns the value of the function.
}
\description{
Necessary for computing covariance function of time-varying fGn
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fgn}
\keyword{simulation}
\keyword{timedep}
