% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_mu_fgn}
\alias{compute_mu_fgn}
\title{Compute mu in C for fGn model}
\usage{
compute_mu_fgn(mu, n, tforcing, Hs)
}
\arguments{
\item{mu}{Vector that will be overwritten with the forced response.}

\item{n}{Integer denoting the dimension of \code{mu}.}

\item{tforcing}{Vector with the shifted and scaled known forcing.}

\item{Hs}{Vector describing the time-dependent Hurst exponent.}
}
\description{
Void function that gives the forced response for the fractional Gaussian noise 
model with time-dependent memory in C.
}
\seealso{
\code{\link{mu.computer}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{C}
\keyword{fgn}
\keyword{forcing}
