% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{fgn_timedep_sim}
\alias{fgn_timedep_sim}
\title{Simulate time dependent fGn series}
\usage{
fgn_timedep_sim(n, sigma = 1, a = 0.6, b = 0.3, Hs = NULL)
}
\arguments{
\item{n}{The length of the simulated time series..}

\item{sigma}{The standard deviation of the innovations.}

\item{a}{Intercept in the evolution of the Hurst exponent.}

\item{b}{Slope in the evolution of the Hurst exponent.}

\item{Hs}{Numeric of length \code{n}. If evolution of Hurst exponent is 
to be given explicitly this is done here. Overrides \code{a} and \code{b}.}
}
\value{
Returns the simulated time series as a \code{numeric} object.
}
\description{
This function produces samples from a given time dependent fractional Gaussian noise process.
}
\examples{
n = 200
sims = fgn_timedep_sim(n,sigma=1,a=0.2,b=0.7)
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{ar1}
\keyword{simulation}
\keyword{timedep}
