% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{fgncor}
\alias{fgncor}
\title{Correlation of time-dependent fractional Brownian motion}
\usage{
fgncor(t, s, Ht, Hs)
}
\arguments{
\item{t}{First index.}

\item{s}{Second index.}

\item{Ht}{Hurst exponent at first index}

\item{Hs}{Hurst exponent at second index}
}
\value{
Returns the correlation between indices 't' and 's' of the time-dependent fBm.
}
\description{
Correlation of fractional Brownian motion with time-dependent increments.
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fbm}
\keyword{simulation}
\keyword{timedep}
