% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{from.theta}
\alias{from.theta}
\title{Transform from internal scaling}
\usage{
from.theta(theta, time_norm, do.forcing = FALSE)
}
\arguments{
\item{theta}{\code{numeric} of parameters in internal scaling.}

\item{time_norm}{\code{numeric} that includes normalized time steps.}

\item{do.forcing}{\code{Boolean} that indicates whether or not forcing is used.}
}
\value{
Returns a list with parameters in user scaling (\code{pars}) as well as 
other useful information obtained from the hyperparameters.
}
\description{
Transforms parameters from internal scaling to user friendly parameters. 
Also includes other information transformed from parameters
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{default}
\keyword{inla.ews}
