% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{gammas}
\alias{gammas}
\title{Adapted gamma function}
\usage{
gammas(a, b, n, t, s)
}
\arguments{
\item{a}{First parameter.}

\item{b}{Second parameter.}

\item{n}{Length of process.}

\item{t}{First index.}

\item{s}{Second index.}
}
\value{
Returns the value of the function.
}
\description{
A version of the gamma function used in other functions.
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fgn}
\keyword{simulation}
\keyword{timedep}
