% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{greensar1}
\alias{greensar1}
\title{Compute Green's function for time-dependent AR(1) model}
\usage{
greensar1(t, s, a, b, n)
}
\arguments{
\item{t}{Row index in the Green's matrix.}

\item{s}{Column index in the Green's matrix.}

\item{a}{Intercept in lienar function for Hurst exponent.}

\item{b}{Slope in lienar function for Hurst exponent.}

\item{n}{Number of data points (used for rescaling \code{b}).}
}
\value{
Returns the evaluation of the Green's function at \code{(t-s)}.
}
\description{
This function computes Green's function.
}
\seealso{
\code{\link{sigmaar1maker}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{Greens}
\keyword{Hurst}
\keyword{exponent}
\keyword{function}
