% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{hyperg2F1}
\alias{hyperg2F1}
\title{Hypergeometric function}
\usage{
hyperg2F1(a, b, c, z, trunc = 300)
}
\arguments{
\item{a}{First parameter.}

\item{b}{Second parameter.}

\item{c}{Third parameter.}

\item{z}{Fourth parameter.}

\item{trunc}{The number of terms used before truncating the infinite series.}
}
\value{
Returns the hypergeometric function evaluated using parameters \code{a,b,c} and \code{z}.
}
\description{
The hypergeometric function 2F1(a,b;c;z).
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fgn}
\keyword{simulation}
\keyword{timedep}
