% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{incBeta}
\alias{incBeta}
\title{Incomplete beta function}
\usage{
incBeta(z, a, b, trunc)
}
\arguments{
\item{z}{The upper bound of the integral of the ordinary beta function.}

\item{a}{First parameter.}

\item{b}{Second parameter.}

\item{trunc}{The number of terms used before truncating the infinite series.}
}
\value{
Returns the value of the incomplete beta function.
}
\description{
The incomplete beta function (allows for some negative parameters).
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fgn}
\keyword{simulation}
\keyword{timedep}
