% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.ews.R
\name{inla.ews}
\alias{inla.ews}
\title{Bayesian analysis of early warning signals}
\usage{
inla.ews(
  data,
  forcing = numeric(0),
  formula = NULL,
  model = "ar1",
  compute.mu = 0,
  timesteps = NULL,
  log.prior = NULL,
  stepsize = 0.005,
  num.threads = 1,
  nsims = 10000,
  do.cgeneric = FALSE,
  inla.options = NULL,
  print.progress = FALSE,
  memory.true = NULL
)
}
\arguments{
\item{data}{Numeric describing the observations from which early warning signals is to be detected.}

\item{forcing}{Numeric describing potential forcing to the system. If no forcing
is included, set to \code{numeric(0)} (default).}

\item{formula}{Formula describing the linear predictor}

\item{model}{Character string describing which noise model should be used.
Currently only supports \code{"ar1"} (default) and \code{"ar2"} (nested AR(1) model).}

\item{compute.mu}{Should the forced response be computed? Set to \code{0} if not, 
\code{1} if only mean and standard deviation should be computed or \code{2} if quantiles
should be computed (this is slower).}

\item{timesteps}{Numeric vector describing time steps of input variables (if steps are not equidistant).}

\item{log.prior}{Function expressing the logarithm of the joint prior for the 
hyperparameters in internal scaling. \code{theta[1]} corresponds to \code{log(kappa)}, 
\code{theta[2]} corresponds to the numerically stable parameterization of \code{b} 
and \code{theta[3]} corresponds to the parameterization of \code{a}. see examples 
for a demonstration. If \code{log.prior=NULL}, default values will be selected. These are
the gamma distribution on \code{kappa} with \code{shape=1} and rate=\code{0.01}, and uniform
priors on \code{a} and \code{b}.}

\item{stepsize}{Numeric stating the step size used in the optimization procedure in 
the INLA program. If convergence cannot be achieved, then changing this will sometimes help. If 
an array of length>=2 is given, INLA will rerun, starting at the previous optima, with given stepsizes. 
Default value is \code{stepsize=0.01}.}

\item{num.threads}{Integer stating how many cores should be used by the INLA program. 
For rgeneric models (as is used here) stability is sometimes improved if \code{num.threads=1}, 
which is the default value here. Increasing this might improve runtime.}

\item{nsims}{Integer stating the number of simulations used in Monte Carlo estimation of parameters
not obtained directly from INLA. Default 10000.}

\item{do.cgeneric}{Boolean stating whether or not the model should be specified using
\code{cgeneric} rather than \code{rgeneric} in cases where this has been implemented. 
Gives improved speed, but bugs are more likely. PC priors are not implemented 
for this model, and any adjustments to the source code will be more difficult.}

\item{inla.options}{List giving options (such as step length and the number of threads to be used),
to the \code{inla}-program. Any absent options are set to the default \code{INLA} options.}

\item{print.progress}{boolean indicating whether or not progress should be printed to screen.}

\item{memory.true}{numeric describing the true memory values. Can be used in 
\code{\link{plot.inla.ews}}.}
}
\value{
Returns an S3 object of class \code{inla.ews}. This object contains 
all results from the \code{inla}-program and associated summary statistics.
}
\description{
Fits a Bayesian hierarchical model with time-dependent correlation using INLA.
}
\examples{
\donttest{
set.seed(123)
n = 1000
sigma = 1
a=0.3
b=0.5
time = 1:n
phis = a+b*seq(0,1,length.out=n)
data=ar1_timedep_sim(n,sigma=sigma,a=a,b=b)


object = inla.ews(data,model="ar1", print.progress=TRUE,
                    memory.true=phis)
summary(object)
plot(object)

## customize prior ##
my.log.prior = function(theta){ #numeric of length 3 (log(kappa), theta_b, theta_a)
  lprior = dnorm(theta[1], sd=1, log=TRUE) + 
           dnorm(theta[2], sd=1, log=TRUE) + 
           dnorm(theta[3], sd=1, log=TRUE)
  return (lprior)
}
object2 = inla.ews(data, log.prior=my.log.prior,model="ar1",compute.mu=FALSE, print.progress=TRUE,
                    memory.true=phis)

summary(object2)
}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{INLA}
\keyword{early}
\keyword{signal}
\keyword{warning}
