% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{inla.options.default}
\alias{inla.options.default}
\title{Default variables in events}
\usage{
inla.options.default()
}
\value{
Returns a list including default values for all variables in \code{inla.options}.
}
\description{
Sets the default variables in the list \code{events} used to specify the climatic
periods and separating events used in the linear predictor. The list contains the following arguments:
\describe{
  \item{\code{num.threads} }{Integer describing the number of cores used by the INLA program. 
  For rgeneric models, stabiltiy is sometimes improved by setting this equal to \code{1} (default).}
  \item{\code{control.inla} }{List containing \code{h=0.01} and \code{restart=1}. 
  \code{h} denotes the step size. Changing this might help reach convergence. 
  \code{restart} is an integer indicating how many times INLA should be restart to 
  at the found optimum. This might help improve the accuracy of the optimization procedure.}
  \item{\code{control.family} }{List with settings to tell INLA to set the variance of the 
  Gaussian likelihood to be fixed and equal to \code{exp(-12)}.}
}
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{default}
\keyword{inla}
\keyword{inla.ews}
\keyword{options}
