% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{mu.computer}
\alias{mu.computer}
\title{Computes the forced response}
\usage{
mu.computer(
  forcing,
  sigmaf,
  F0,
  a = 0.6,
  b = 0.2,
  memory = NULL,
  model = "ar1"
)
}
\arguments{
\item{forcing}{Known forcing.}

\item{sigmaf}{Scaling parameter of forcing.}

\item{F0}{Shift parameter of forcing.}

\item{a}{Intercept of memory evolution.}

\item{b}{Slope of memory evolution.}

\item{memory}{If memory is to be given explicitly this is done here as \code{numeric} object.}

\item{model}{Character string describing which model is assumed for the noise. 
Can be \code{"ar1"} (default) or \code{"fgn"}.}
}
\value{
Returns the forced response based on the input arguments.
}
\description{
This function computes the forced response given known forcing and memory development.
}
\examples{
n = 300
F0 = -1
sigmaf = 0.2
a = 0.6
b = 0.3/n
time = 1:n
Hs = a+b*time
forcing = arima.sim(model=list(ar=c(0.95)),n=n,sd=sqrt(1-0.95^2))+1:n/n*0.2
zz = sigmaf*(F0+forcing)

muvek = mu.computer(forcing,sigmaf,F0,memory=Hs,model="fgn")

}
\seealso{
\code{\link{inla.ews}, \link{forcingmaker}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{forcing}
\keyword{mean}
\keyword{response}
