% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{mucomputer}
\alias{mucomputer}
\title{Compute forcing response}
\usage{
mucomputer(pars, forcing, time_norm, as.theta = FALSE)
}
\arguments{
\item{pars}{\code{numeric} of parameters.}

\item{forcing}{\code{numeric} that includes known forcing component.}

\item{time_norm}{\code{numeric} that includes normalized time steps.}

\item{as.theta}{\code{Boolean} that indicates whether or not \code{pars} is given 
in internal or user scaling.}
}
\value{
Returns a \code{numeric} with the estimated forcing response.
}
\description{
Computes forcing response given parameters in either internal or user scaling using Monte Carlo sampling.
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{default}
\keyword{inla.ews}
