% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inla.ews.R
\name{print.inla.ews}
\alias{print.inla.ews}
\title{Print function for a \code{summary.inla.ews} object}
\usage{
\method{print}{inla.ews}(x, digits = 4L, ...)
}
\arguments{
\item{x}{\code{inla.ews} S3 class. Output of \code{\link{inla.ews}} function.}

\item{digits}{Integer indicating the number of decimal places to print.}

\item{...}{Other arguments.}
}
\description{
Displays the return object from \code{summary.inla.ews} to an easily 
readable format.
}
\examples{
\donttest{
n = 500
sigma = 1
a=0.2
b=0.7/n
time = 1:n
phis = a+b*time
s=numeric(n)
s[1] = rnorm(1,mean=0,sd=sigma)
for(i in 2:n){
  s[i] = rnorm(1, mean=phis[i]*s[i-1],sd=sigma)
}

object = inla.ews(data=s,model="ar1", memory.true=phis)
print(object)
}
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{inla.ews}
\keyword{print}
