% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.inla.ews.R
\name{print.summary.inla.ews}
\alias{print.summary.inla.ews}
\title{Print function for the \code{inla.ews} object}
\usage{
\method{print}{summary.inla.ews}(x, digits = 4L, ...)
}
\arguments{
\item{x}{\code{summary.inla.ews} S3 class. Output of \code{\link{summary.inla.ews}} function.}

\item{digits}{Integer indicating the number of decimal places to be used.}

\item{...}{Other arguments.}
}
\description{
Displays the return object from \code{summary.inla.ews} to an easily readable format.
}
\examples{
\donttest{
n = 200
sigma = 1
a=0.2
b=0.7/n
time = 1:n
phis = a+b*time
data=ar1_timedep_sim(n,phis=phis)

object = inla.ews(data,model="ar1", print.progress=TRUE,
                    memory.true=phis)
summaryobj = summary(object)
print(summaryobj)
}
}
\seealso{
\code{\link{inla.ews},\link{summary.inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{inla.ews}
\keyword{print.summary}
