% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{R2test}
\alias{R2test}
\title{Covariance of time-dependent fractional Brownian motion}
\usage{
R2test(a, b, n, t, s)
}
\arguments{
\item{a}{Starting point of Hurst exponent (intercept).}

\item{b}{Slope of Hurst exponent.}

\item{n}{Length of process.}

\item{t}{First index.}

\item{s}{Second index.}
}
\value{
Returns the covariance function of the time-dependent fBm.
}
\description{
Covariance function of fractional Brownian motion with time-dependent increments.
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fbm}
\keyword{simulation}
\keyword{timedep}
