% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultgather.R
\name{resultgather}
\alias{resultgather}
\title{Gather results from INLA fit
This function gathers and formats results from INLA fit.}
\usage{
resultgather(object, nsims = 10000, print.progress)
}
\arguments{
\item{object}{S3 object of type \code{inla.ews} which includes result from 
\code{inla}-program.}

\item{nsims}{Integer stating the number of simulations used in Monte Carlo estimation of parameters
not obtained directly from INLA. Default 10000.}

\item{print.progress}{boolean indicating if progress should be printed to screen.}
}
\value{
Returns the same S3 object of class \code{inla.ews} as included in the 
input arguments, but appends summary statistics and Monte Carlo simulations in 
\code{object\$results}.
}
\description{
Gather results from INLA fit
This function gathers and formats results from INLA fit.
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{INLA}
\keyword{early}
\keyword{signal}
\keyword{summary}
\keyword{warning}
