% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{Rfgn}
\alias{Rfgn}
\title{Covariance of time-dependent fractional Gaussian noise}
\usage{
Rfgn(a, b, n, t, s)
}
\arguments{
\item{a}{Starting point of Hurst exponent (intercept).}

\item{b}{Slope of Hurst exponent.}

\item{n}{Length of process.}

\item{t}{First index.}

\item{s}{Second index.}
}
\value{
Returns the covariance function of the time-dependent fGn.
}
\description{
Covariance function of fractional Gaussian noise with time-dependent increments.
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{fgn}
\keyword{simulation}
\keyword{timedep}
