% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeneric.ar2.R
\name{rgeneric.ar2.forcing}
\alias{rgeneric.ar2.forcing}
\title{rgeneric model for AR(2) time-dependent process}
\usage{
rgeneric.ar2.forcing(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\arguments{
\item{cmd}{Vector containing list of function names necessary for the rgeneric model.}

\item{theta}{Vector describing the hyperparameters in internal scaling.}
}
\description{
Defines the rgeneric model structure for the AR(2)/nested AR(1) model with time-dependent
correlation. This includes key functions that provide information on precision matrix, mean vector, priors, graph etc.
This is intended for internal use only, but the documentation is included here in case someone want to change something.
}
