% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{sigmaar1maker}
\alias{sigmaar1maker}
\title{Construct covariance matrix of time-dependent ar1g model}
\usage{
sigmaar1maker(sigma, a, b, n)
}
\arguments{
\item{sigma}{Numeric of length 1. The innovation standard deviation}

\item{a}{Numeric of length 1. Intercept of the memory parameter}

\item{b}{Numeric of length 1. Slope of the memory parameter}

\item{n}{Integer. Length of the simulation.}
}
\value{
Returns the covariance matrix of the time dependent fGn process. This is 
a dense \code{matrix} object.
\code{object\$results}.
}
\description{
This function constructs the covariance matrix of the ar1g process under the 
assumption that the Hurst exponent changes over time. In the context of the 
\code{inla.ews} package this function is only used for simulating examples.
}
\examples{
n=50
a=0.6
b=0.2
sigma = 1
cov.matrix <- sigmaar1maker(sigma,a,b,n)
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{covariance}
\keyword{fgn}
\keyword{matrix}
