% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{to.theta}
\alias{to.theta}
\title{Transform to internal scaling}
\usage{
to.theta(pars, time_norm, do.forcing = FALSE)
}
\arguments{
\item{pars}{\code{numeric} of parameters in user scaling.}

\item{time_norm}{\code{numeric} that includes normalized time steps.}

\item{do.forcing}{\code{Boolean} that indicates whether or not forcing is used.}
}
\value{
Returns a list with parameters in internal scaling (\code{theta}) as well as 
other useful information obtained from the hyperparameters.
}
\description{
Transforms parameters to internal scaling from user friendly parameters. 
Also includes other information transformed from parameters
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{default}
\keyword{inla.ews}
