% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.ews.R
\name{inla.ews}
\alias{inla.ews}
\title{Bayesian analysis of early warning signals}
\usage{
inla.ews(
  data,
  forcing = numeric(0),
  formula = NULL,
  model = "ar1",
  compute.mu = 0,
  timesteps = NULL,
  stepsize = 0.01,
  num.threads = 1,
  inla.options = NULL,
  print.progress = FALSE,
  memory.true = NULL
)
}
\arguments{
\item{data}{Numeric describing the observations from which early warning signals is to be detected.}

\item{forcing}{Numeric describing potential forcing to the system. If no forcing
is included, set to \code{numeric(0)} (default).}

\item{formula}{Formula describing the linear predictor (NOT YET IMPLEMENTED)}

\item{model}{Character string describing which noise model should be used.
Currently supports \code{"ar1"} (default) and \code{"fgn"}.}

\item{compute.mu}{Should the forced response be computed? Set to \code{0} if not, 
\code{1} if only mean and standard deviation should be computed or \code{2} if quantiles
should be computed (this is slower).}

\item{timesteps}{Numeric vector describing time steps of input variables (if steps are not equidistant).}

\item{stepsize}{Numeric stating the step size used in the optimization procedure in 
the INLA program. If convergence cannot be achieved, then changing this will sometimes help. 
Default value is \code{stepsize=0.01}.}

\item{num.threads}{Integer stating how many cores should be used by the INLA program. 
For rgeneric models (as is used here) stability is sometimes improved if \code{num.threads=1}, 
which is the default value here. Increasing this might improve runtime.}

\item{inla.options}{List giving options (such as step length and the number of threads to be used),
to the \code{inla}-program. Any absent options are set to the default \code{INLA} options.}

\item{print.progress}{boolean indicating whether or not progress should be printed to screen.}

\item{memory.true}{numeric describing the true memory values. Can be used in 
\code{\link{plot.inla.ews}}.}
}
\value{
Returns an S3 object of class \code{inla.ews}. This object contains 
all results from the \code{inla}-program and associated summary statistics.
}
\description{
Fits a Bayesian hierarchical model with time-dependent correlation using INLA.
}
\examples{
\donttest{
### AR(1) simulation example ###
set.seed(123)
n = 300
sigma = 1
a=0.2
b=0.7/n
F0 = 3
sigmaf = 0.3
time = 1:n
phis = a+b*time
noise=ar1_timedep_sim(n,phis=phis)

forcing = arima.sim(model=list(ar=c(0.95)),n=n,sd=sqrt(1-0.95^2))+1:n/n
zz = sigmaf*(F0+forcing)

lambdas = phis-1
muvek = mu.computer(forcing,sigmaf,F0,memory=phis,model="ar1")

data = noise + muvek

object = inla.ews(data,forcing,model="ar1",compute.mu=2, print.progress=TRUE,
                    memory.true=phis)
summary(object)
plot(object)


### fGn simulation example ###
set.seed(123)
n=200
sigma = 1.2
time=1:n
a = 0.6
b = 0.35/n
Hs = a+b*time

data = fgn_timedep_sim(n,Hs=Hs)

object = inla.ews(data,model="fgn",memory.true=Hs)
summary(object)
plot(object)


}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{INLA}
\keyword{early}
\keyword{signal}
\keyword{warning}
