% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.ews.R
\name{inla.ews}
\alias{inla.ews}
\title{Bayesian analysis of early warning signals}
\usage{
inla.ews(
  data,
  forcing = numeric(0),
  formula = NULL,
  model = "ar1",
  inla.options = NULL,
  print.progress = FALSE,
  memory.true = NULL
)
}
\arguments{
\item{data}{Numeric describing the observations from which early warning signals is to be detected.}

\item{forcing}{Numeric describing potential forcing to the system. If no forcing
is included, set to \code{numeric(0)} (default).}

\item{formula}{Formula describing the linear predictor (NOT YET IMPLEMENTED)}

\item{model}{Character string describing which noise model should be used.
Currently supports \code{"ar1"} (default) and \code{"fgn"}.}

\item{inla.options}{List giving options (such as step length and the number of threads to be used),
to the \code{inla}-program. Any absent options are set to the default \code{INLA} options.}

\item{print.progress}{boolean indicating whether or not progress should be printed to screen.}

\item{memory.true}{numeric describing the true memory values. Can be used in 
\code{\link{plot.inla.ews}}.}
}
\value{
Returns an S3 object of class \code{inla.ews}. This object contains 
all results from the \code{inla}-program and associated summary statistics.
}
\description{
Fits a Bayesian hierarchical model with time-dependent correlation using INLA.
}
\examples{
\donttest{
### AR(1) simulation example ###
n = 300
sigma = 1
a=0.2
b=0.7/n
F0 = 3
sigmaf = 2
time = 1:n
phis = a+b*time
s=numeric(n)
s[1] = rnorm(1,mean=0,sd=sigma)
for(i in 2:n){
  s[i] = rnorm(1, mean=phis[i]*s[i-1],sd=sigma)
}

forcing = arima.sim(model=list(ar=c(0.95)),n=n,sd=sqrt(1-0.95^2))+1:n/n*7
zz = sigmaf*(F0+forcing)

lambdas = phis-1

struct = exp(lambdas*(1:n)-0.5)
muvek=numeric(n)
for(i in 1:n){
  muvek[i] = rev(struct[1:i])\%*\%zz[1:i]
}
data = s + muvek

object = inla.ews(data,forcing,model="ar1", print.progress=TRUE,
                    memory.true=phis)
summary(object)
plot(object)


### fGn simulation example ###
n=200
sigma = 1.2
time=1:n
a = 0.6
b = 0.35/n
Hs = a+b*time

sigmat = sigmaHmaker(sigma,Hs)
sigmachol = chol(sigmat)
data = sigmachol\%*\%rnorm(n)

object = inla.ews(data,model="fgn",memory.true=Hs)
summary(object)
plot(object)

}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{INLA}
\keyword{early}
\keyword{signal}
\keyword{warning}
