% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{sigmaHmaker}
\alias{sigmaHmaker}
\title{Construct covariance matrix of time-dependent fGn model}
\usage{
sigmaHmaker(sigma, Hs)
}
\arguments{
\item{sigma}{Numeric of length 1. The innovation standard deviation}

\item{Hs}{Numeric of length corresponding to the length of the process. This gives 
the value of the Hurst exponent at each time point.}
}
\value{
Returns the covariance matrix of the time dependent fGn process. This is 
a dense \code{matrix} object.
\code{object\$results}.
}
\description{
This function constructs the covariance matrix of the fGn process under the 
assumption that the Hurst exponent changes over time. In the context of the 
\code{inla.ews} package this function is only used for simulating examples.
}
\examples{
n=100
Hs = seq(from=0.6,to=0.8,length.out=n)
sigma = 1
cov.matrix <- sigmaHmaker(sigma,Hs)
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{covariance}
\keyword{fgn}
\keyword{matrix}
