% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.inla.ews.R
\name{summary.inla.ews}
\alias{summary.inla.ews}
\title{Summary-function for \code{inla.ews} object}
\usage{
\method{summary}{inla.ews}(object, digits = 4L, ...)
}
\arguments{
\item{object}{\code{inla.ews} S3 class. Output of \code{\link{inla.ews}} function.}

\item{digits}{Integer indicating the number of decimal places to be used.}

\item{...}{Other arguments.}
}
\value{
Returns an object of class \code{summary.inla.ews}.
}
\description{
Displays a clear summary for the information stored in the \code{inla.ews} S3 object 
returned by the \code{\link{inla.ews}} function.
}
\examples{
\donttest{
n = 500
sigma = 1
a=0.2
b=0.7/n
time = 1:n
phis = a+b*time
s=numeric(n)
s[1] = rnorm(1,mean=0,sd=sigma)
for(i in 2:n){
  s[i] = rnorm(1, mean=phis[i]*s[i-1],sd=sigma)
}

object = inla.ews(data=s,model="ar1", memory.true=phis)
summary(object)
}
}
\seealso{
\code{\link{inla.ews}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{inla.ews}
\keyword{summary}
